+++ src/pages/MainPage/MainPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ShopDescription from '../../components/MainPage/ShopDescription';
+import ShopDescription from '../../components/MainPage/ShopDescription/ShopDescription';
 
 function MainPage() {
   return <ShopDescription />;
